pkg/proc: More file organization
